+++ example/integration_test/sc-IL-internalLimits/IL_200_key_length_truncate_test.dart
@@ -67,12 +67,12 @@ void main() {
       } else if (a == 5) {
         // 0) Custom Event
         var eventRaw = json.decode(queryParams['events']![0]);
-        Map<String, dynamic> event = Platform.isIOS ? eventRaw[0] : eventRaw;
+        Map<String, dynamic> event = eventRaw[0];
         expect(event['key'], 'Event With Sum And Segment'.substring(0, MAX_KEY_LENGTH));
         expect(event['segmentation']['Country'.substring(0, MAX_KEY_LENGTH)], 'Turkey');
         expect(event['segmentation']['Age'.substring(0, MAX_KEY_LENGTH)], '28884');
         // 1) View Start (legacy)
-        Map<String, dynamic> view = Platform.isIOS ? eventRaw[1] : json.decode(queryParams['events']![1]);
+        Map<String, dynamic> view = eventRaw[1];
         expect(view['key'], '[CLY]_view');
         expect(view['segmentation']['Cats'.substring(0, MAX_KEY_LENGTH)], '12345');
         // merges with moose if limit is small
@@ -85,14 +85,14 @@ void main() {
         expect(view['segmentation']['NotCamel'.substring(0, MAX_KEY_LENGTH)], 'Deerz');
         expect(view['segmentation']['Moose'.substring(0, MAX_KEY_LENGTH)], 'Deer');
         // 2) View End (legacy)
-        view = Platform.isIOS ? eventRaw[2] : json.decode(queryParams['events']![2]);
+        view = eventRaw[2];
         expect(view['key'], '[CLY]_view');
         expect(view['segmentation']['segment'], Platform.isIOS ? 'iOS' : 'Android');
         expect(view['segmentation']['name'], 'HomePage'.substring(0, MAX_KEY_LENGTH));
         expect(view['segmentation']['Camel'.substring(0, MAX_KEY_LENGTH)], 666);
         expect(view['segmentation']['NotCamel'.substring(0, MAX_KEY_LENGTH)], 'Deerz');
         // 3) View Start (AutoStopped)
-        view = Platform.isIOS ? eventRaw[3] : json.decode(queryParams['events']![3]);
+        view = eventRaw[3];
         expect(view['key'], '[CLY]_view');
         expect(view['segmentation']['Cats'.substring(0, MAX_KEY_LENGTH)], 12345);
         // expect(view['segmentation']['Moons'.substring(0, MAX_KEY_LENGTH)], 9.9866); // merges with moose
@@ -102,15 +102,6 @@ void main() {
         expect(view['segmentation']['visit'], Platform.isIOS ? 1 : '1');
         expect(view['segmentation']['NotCamel'.substring(0, MAX_KEY_LENGTH)], 'Deerz');
         expect(view['segmentation']['Moose'.substring(0, MAX_KEY_LENGTH)], 'Deer');
-        // 4) View End (AutoStopped)
-        if (Platform.isAndroid) {
-          view = json.decode(queryParams['events']![4]);
-          expect(view['key'], '[CLY]_view');
-          expect(view['segmentation']['segment'], Platform.isIOS ? 'iOS' : 'Android');
-          expect(view['segmentation']['name'], 'hawk'.substring(0, MAX_KEY_LENGTH));
-          expect(view['segmentation']['Camel'.substring(0, MAX_KEY_LENGTH)], 666);
-          expect(view['segmentation']['NotCamel'.substring(0, MAX_KEY_LENGTH)], 'Deerz');
-        }
       } else if (a == 6) {
         Map<String, dynamic> userDetails = json.decode(queryParams['user_details']![0]);
         expect(userDetails['custom']['special_value'.substring(0, MAX_KEY_LENGTH)], 'something special');