+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 24.4.1
 * Mitigated an issue where user data operation `pull` won't work
 * Mitigated an issue where custom apm metrics were cast to string
+* Mitigated an issue where user property `byear` was cast to string
 
 * Updated underlying Android SDK version to 24.4.1
 * Updated underlying iOS SDK version to 24.4.1